requirements.txt
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pycparser==2.23
PyMuPDF==1.26.6
pypdfium2==5.0.0
pytesseract==0.3.13
